FastFermionSampling

Documentation for FastFermionSampling.

FastFermionSampling.AHmodel

FastFermionSampling.AHmodel

FastFermionSampling.Gutzwiller

FastFermionSampling.LatticeRectangular

FastFermionSampling.FFS

FastFermionSampling.fast_G_update

FastFermionSampling.fast_update

FastFermionSampling.getHmat

FastFermionSampling.getOL

FastFermionSampling.getxprime

FastFermionSampling.AHmodel — Type

Anderson-Hubbard Model

lattice: LatticeRectangular{B} The lattice structure t: Float64 Hopping parameter W: Float64 Disorder strength U: Float64 On-site interaction strength Nup: Int Number of up spins Ndown: Int Number of down spins omega: Vector{Float64} Random on-site energies Uup: Matrix Unitary matrix for up spins Udown: Matrix Unitary matrix for down spins

FastFermionSampling.AHmodel — Method

generate Anderson-Hubbard model and get the sampling ensemble

FastFermionSampling.Gutzwiller — Method

add Gutzwiller Ansatz where G = exp(-g/2 ∑i (ni - nmean)^2), ΨG = G Ψ_0

FastFermionSampling.LatticeRectangular — Method

LatticeRectangular(nx::Int, ny::Int, B::Periodic)
LatticeRectangular(nx::Int, ny::Int, B::Open)

FastFermionSampling.FFS — Method

Employing Fast Fermion Sampling Algorithm to sample free Fermions

FFS([rng=default_rng()], U::AbstractMatrix)

U: the sampling ensemble, a matrix of size L x N, where L is the number of energy states and N is the number of Fermions

FastFermionSampling.fast_G_update — Method

Fast Gutzwiller Factor update technique from Becca and Sorella 2017 Should input whole conf

FastFermionSampling.fast_update — Method

Fast computing technique from Becca and Sorella 2017

FastFermionSampling.getHmat — Method

Get the non-interacting Anderson model Hamiltonian Matrix to construct HF states

FastFermionSampling.getOL — Method

The observable OL = <x|H|ΨG>/<x|ΨG>

FastFermionSampling.getxprime — Method

return |x'> = H|x> where H = -t ∑<i,j> ci^† cj + U ∑i ni↓ ni↑ + ∑i ωi n_i