FastFermionSampling

Documentation for FastFermionSampling.

FastFermionSampling.AHmodel

FastFermionSampling.AHmodel

FastFermionSampling.Gutzwiller

FastFermionSampling.LatticeRectangular

FastFermionSampling.MC

Carlo.init!

Carlo.register_evaluables

FastFermionSampling.FFS

FastFermionSampling.fast_G_update

FastFermionSampling.fast_update

FastFermionSampling.getHmat

FastFermionSampling.getOL

FastFermionSampling.getOg

FastFermionSampling.getxprime

FastFermionSampling.AHmodel — Type

Anderson-Hubbard Model

lattice : LatticeRectangular{B} The lattice structure

t: Float64 Hopping parameter

W : Float64 Disorder strength, on site energy is sampled from N(0, W/2)

U : Float64 On-site interaction strength

Nup : Int Number of up spins

Ndown : Int Number of down spins

omega : Vector{Float64} Random on-site energies

U_up : Matrix{Float64} Unitary matrix for up spins

U_down : Matrix{Float64} Unitary matrix for down spins

FastFermionSampling.AHmodel — Method

AHmodel(lattice::LatticeRectangular{B}, t::Float64, W::Float64, U::Float64, N_up::Int, N_down::Int)

Generate Anderson-Hubbard model and get the sampling ensemble

FastFermionSampling.Gutzwiller — Type

Gutzwiller Ansatz

g : Float64 Gutzwiller factor

Store the Gutzwiller factor parameter g.

The Gutzwiller factor is defined as $G = \exp(-g/2 \sum_i (n_i - n_{mean})^2)$.

FastFermionSampling.LatticeRectangular — Method

LatticeRectangular(nx::Int, ny::Int, B::Periodic)
LatticeRectangular(nx::Int, ny::Int, B::Open)

Generate a rectangular lattice with periodic or open boundary conditions

nx : number of sites in x-direction

ny : number of sites in y-direction

B : boundary condition, either Periodic or Open

FastFermionSampling.MC — Method

MC(params::AbstractDict)

Create a Monte Carlo object

Carlo.init! — Method

Carlo.init!(mc::MC, ctx::MCContext, params::AbstractDict)

Initialize the Monte Carlo object params

nx : Int number of sites in x direction

ny : Int number of sites in y direction

B : AbstractBoundary boundary condition, Periodic or Open

t : Float64 hopping parameter

W : Float64 disorder strength

U : Float64 on-site interaction strength

N_up : Int number of up spins

N_down : Int number of down spins

Carlo.register_evaluables — Method

fg the gradient of ⟨E_g⟩

Get the gradient of the observable, $f_g = - ∂ ⟨E_g⟩/ ∂ g$.

`` \begin{align} fk &= -2 ℜ[⟨OL(x)^* × (Og(x)- ⟨Og⟩) ⟩ ]\ &= -2 ℜ[⟨OL(x)^* × Og(x) ⟩ - ⟨OL(x)^* ⟩ × ⟨Og⟩ ]\ \end{align} ``

fisherScalar

Get the Fisher Matrix of the observable, $S_{k,k'} = ℜ⟨⟨O_k O_{k'}⟩⟩ = ℜ( ⟨O_k O_{k'}⟩ -⟨O_k⟩ ⟨O_{k'}⟩ )$ where $k$ and $k'$ are labels of the parameters of the model.

When ansatz has only one parameter, the Fisher Matrix is a scalar, and the Fisher Information is the inverse of the Fisher Matrix.

** Structure Factor ** in low momentum

$N_q = ⟨⟨n_qn_{-q}⟩⟩_{disorder}- ⟨⟨n_q⟩⟨n_{-q}⟩⟩_{disorder}$

FastFermionSampling.FFS — Method

FFS([rng=default_rng()], U::AbstractMatrix)

Employing Fast Fermion Sampling Algorithm to sample free Fermions

U : the sampling ensemble, a matrix of size L x N, where L is the number of energy states and N is the number of Fermions

FastFermionSampling.fast_G_update — Method

fast_G_update(newwholeconf::BitStr{N,T}, oldwholeconf::BitStr{N,T}, g::Float64, n_mean::Float64) where {N,T}

Fast Gutzwiller Factor update technique from Becca and Sorella 2017

Should input whole configuration

FastFermionSampling.fast_update — Method

fast_update(U::AbstractMatrix, Uinvs::AbstractMatrix, newconf::BitStr{N,T}, oldconf::BitStr{N,T}) where {N,T}

Fast computing technique from Becca and Sorella 2017

FastFermionSampling.getHmat — Method

getHmat(lattice::LatticeRectangular{B}, t::Float64, omega::Vector{Float64}, N_up::Int, N_down::Int)'

Get the non-interacting Anderson model Hamiltonian Matrix to construct Slater Determinants

FastFermionSampling.getOL — Method

getOL(orb::AHmodel, conf_up::BitVector, conf_down::BitVector, g::Float64)

The observable $O_L = \frac{<x|H|\psi_G>}{<x|\psi_G>}$

FastFermionSampling.getOg — Method

getOg(orbitals::AHmodel{B}, conf_up::BitVector, conf_down::BitVector)

The local operator to update the variational parameter g $mathcal{O}_k(x)=\frac{\partial \ln \Psi_\alpha(x)}{\partial \alpha_k}$

FastFermionSampling.getxprime — Method

getxprime(orb::AHmodel{B}, x::BitStr{N,T}) where {B,N,T}

return $|x'> = H|x>$ where $H = -t ∑_{<i,j>} c_i^† c_j + U ∑_i n_{i↓} n_{i↑} + ∑_i ω_i n_i$